+++ repo_people/__init__.py
@@ -1,1 +1,1 @@
-from .repo_people import *
+from .repo_people import RepoPeople, UserDataView